Refactor code: improve packaging, extract methods & fields
