+++ scrapper/src/main/java/com/scrappy/scrapper/html/api/HtmlScrapper.java
@@ -1,4 +1,4 @@
-package com.scrappy.scrapper.html;
+package com.scrappy.scrapper.html.api;
 
 import com.scrappy.scrapper.common.Book;
