+++ scrapper/src/main/java/com/scrappy/scrapper/html/core/niedziela/NiedzielaScrapper.java
@@ -1,8 +1,10 @@
-package com.scrappy.scrapper.html;
+package com.scrappy.scrapper.html.core.niedziela;
 
 import com.scrappy.scrapper.common.Book;
+import com.scrappy.scrapper.html.api.HtmlScrapper;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
 import java.io.IOException;
@@ -15,6 +17,10 @@ import java.util.regex.Pattern;
 import static java.util.Collections.emptyList;
 
 class NiedzielaScrapper implements HtmlScrapper {
+  
+  private static final String BASE_URL = "https://ksiegarnia.niedziela.pl/";
+  
+  private static final String BOOKSTORE = "Księgarnia Niedziela";
   
   protected Document retrievePromoBooks() throws IOException {
     return Jsoup.connect("https://ksiegarnia.niedziela.pl/site/promocje/").get();
@@ -38,16 +44,15 @@ class NiedzielaScrapper implements HtmlScrapper {
       final List<Book> books = new ArrayList<>();
       final Document doc = this.retrievePromoBooks();
       final Elements elements = doc.getElementsByClass("polecamy");
-      elements.forEach(element -> {
-        String bookUrl = "https://ksiegarnia.niedziela.pl/" + element.getElementsByTag("a").attr("href");
+      elements.forEach(e -> {
         final Book book = Book.builder()
-                        .setTitle(element.getElementsByClass("polecamy_tytul").text())
-                        .setAuthor(element.getElementsByClass("polecamy_dzial").text())
-                        .setListPrice(retrievePriceFrom(element.getElementsByTag("strike").text()))
-                        .setDiscountPrice(retrievePriceFrom(element.getElementsByClass("polecamy_cena").tagName("span").get(0).text()))
-                        .setBookstore("Księgarnia Niedziela")
-                        .setUrl(bookUrl)
-                        .setIsbn(retrieveIsbnFrom(bookUrl)).build();
+                        .setTitle(retrieveTitleFrom(e))
+                        .setAuthor(retrieveAuthorFrom(e))
+                        .setListPrice(retrieveListPriceFrom(e))
+                        .setDiscountPrice(retrieveDiscountPriceFrom(e))
+                        .setBookstore(BOOKSTORE)
+                        .setUrl(retrieveUrlFrom(e))
+                        .setIsbn(retrieveIsbnFrom(retrieveUrlFrom(e))).build();
         books.add(book);
       });
       return books;
@@ -55,6 +60,14 @@ class NiedzielaScrapper implements HtmlScrapper {
       e.printStackTrace();
     }
     return emptyList();
+  }
+  
+  private String retrieveAuthorFrom(final Element element) {
+    return element.getElementsByClass("polecamy_dzial").text();
+  }
+  
+  private String retrieveTitleFrom(final Element element) {
+    return element.getElementsByClass("polecamy_tytul").text();
   }
   
   private String retrieveIsbnFrom(String url) {
@@ -71,6 +84,20 @@ class NiedzielaScrapper implements HtmlScrapper {
       e.printStackTrace();
     }
     return "";
+  }
+  
+  private String retrieveUrlFrom(Element element) {
+    return BASE_URL + element.getElementsByTag("a").attr("href");
+  }
+  
+  private BigDecimal retrieveListPriceFrom(Element element) {
+    String pattern = element.getElementsByTag("strike").text();
+    return retrievePriceFrom(pattern);
+  }
+  
+  private BigDecimal retrieveDiscountPriceFrom(Element element) {
+    String pattern = element.getElementsByClass("polecamy_cena").tagName("span").get(0).text();
+    return retrievePriceFrom(pattern);
   }
   
   private BigDecimal retrievePriceFrom(String pattern) {
